requirements.txt
requests
feedparser
redis
flask
python-dotenv
celery[redis]
gunicorn
